Refactor and fix sass codes
