+++ src/sass/abstracts/mixins/_mixins.scss
@@ -1,0 +1,6 @@
+// Center flexbox.
+@mixin flex-center {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
