+++ src/sass/abstracts/variables/_colors.scss
@@ -1,0 +1,3 @@
+$color__border-card: #ccc;
+$color__shadow-card: rgba(#000, 0.05);
+$color__background-posts: #fafafc;
