+++ src/sass/abstracts/variables/_typography.scss
@@ -1,0 +1,6 @@
+$font__size-card-heading: 2.3rem;
+$font__size-card-emoji-flag: 12.5rem;
+
+$font__size-posts-heading: 2.2rem;
+$font__size-posts-title: 2.1rem;
+$font__size-posts-excerpt: 1.1rem;
