+++ src/sass/commons/_index.scss
@@ -1,0 +1,1 @@
+@forward "wp-block";
