+++ src/sass/commons/_wp-block.scss
@@ -1,0 +1,8 @@
+.wp-block-xwp-country-card {
+
+	&.alignnone,
+	&.alignleft,
+	&.alignright {
+		max-width: 100%;
+	}
+}
