+++ src/sass/components/_index.scss
@@ -1,0 +1,3 @@
+@forward "country-card";
+@forward "loader";
+@forward "related-posts";
