+++ src/sass/editor.scss
@@ -1,3 +1,6 @@
+@use "commons/wp-block";
+@use "components/loader";
+
 .wp-block-xwp-country-card .components-form-token-field__suggestion {
 	white-space: pre-wrap;
 }
