+++ src/sass/style.scss
@@ -1,0 +1,2 @@
+@use "commons/wp-block";
+@use "components";
