--- src/style.scss
@@ -1,69 +1,0 @@
-.xwp-country-card {
-	border: 1px solid #dcdcde;
-	box-shadow: 0px 1px 1px rgba(#000, 0.05);
-
-	.xwp-country-card__media {
-		align-items: center;
-		border-bottom: 1px solid #dcdcde;
-		display: flex;
-		height: calc(200px + 5vw);
-		justify-content: center;
-		overflow: hidden;
-		position: relative;
-		margin: 0px;
-
-		&:before {
-			align-items: center;
-			content: attr(data-emoji-flag);
-			display: flex;
-			filter: blur(3vh) saturate(3) contrast(1) opacity(0.4);
-			font-size: calc(800px + 3vh);
-			height: 100%;
-			justify-content: center;
-			left: 0px;
-			pointer-events: none;
-			position: absolute;
-			top: 0px;
-			width: 100%;
-		}
-	}
-}
-
-.xwp-country-card-flag {
-	font-size: calc(200px + 3vw);
-	line-height: 1;
-}
-
-.xwp-country-card__heading {
-	padding: 1.5rem;
-	margin: 0;
-	font-size: 2rem;
-}
-
-.xwp-country-card__related-posts {
-	border-top: 1px solid #dcdcde;
-	background-color: #fafafc;
-	padding: 1.5rem;
-}
-
-.xwp-country-card__related-posts__heading {
-	margin: 0px;
-	font-size: 1.5rem;
-}
-
-.xwp-country-card__related-posts-list {
-	padding-left: 1.5rem;
-}
-
-.related-post + .related-post {
-	margin-top: 1.5rem;
-}
-
-.related-post .title {
-	margin: 0px;
-	font-size: 1.5rem;
-}
-
-.related-post .excerpt {
-	margin: 0px;
-}
